+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 2.3.1 (Unreleased)
+**NOTICE**
+- Due to data integrity issues, random write operations has been disabled in block cache. Refer [#1484](https://github.com/Azure/azure-storage-fuse/pull/1484) for blocked scenarios.
+
 **Bug Fixes**
 - Fixed the case where file creation using SAS on HNS accounts was returning back wrong error code.
 - [#1402](https://github.com/Azure/azure-storage-fuse/issues/1402) Fixed proxy URL parsing.
@@ -6,9 +9,8 @@
 - In flush operation, the blocks will be committed only if the handle is dirty.
 - Reset block data to null before reuse.
 - Sparse file data integrity issues fixed.
-- Fixed block-cache read of small files where file size is not multiple of kernel buffer size. 
-
-**Features**
+- Fixed block-cache read of small files where file size is not multiple of kernel buffer size.
+- Fixed race condition in random write where a block is being uploaded and written to in parallel.
 
 **Other Changes**
 - LFU policy in file cache has been deprecated.